refactor(Topology): :recycle: Re-organize Graph structure
